Finished update to the current cuwo version
Instead of creating own entities the entity methods (including hostility
setting) are injected into cuwo's entity class. Also added a possibility
to pass the particle data on particle effect creation using the factory
and removed the Color class as a 4-tupel is used now in the packet
structure. Removed a bunch of constants as they are now defined within
cuwo.
Also updated the UML diagram.